ci: only publish helm-chart once (#1045)
Each file found with "release.extra_files" will become an artifact.
Those artifacts cannot be filtered by "publishers[].ids". If the
"release" step does not run (e.g. snapshot), the publishers will not
have any artifacts to publish.

This make the "release.extra_files" glob pattern more strict, and ensure
that the helm chart publisher only runs once, for the helm chart
package.

+++ scripts/publish-helm-chart.sh
@@ -13,6 +13,12 @@ if [[ -z "$CHART_FILE" ]]; then
   exit 1
 fi
 
+# Ensures we only publish once, when $1 == $GORELEASER_ARTIFACT_FILE
+if [[ -n "${GORELEASER_ARTIFACT_FILE}" && "${GORELEASER_ARTIFACT_FILE}" != "${CHART_FILE}" ]]; then
+  echo "skipping artifact: ${GORELEASER_ARTIFACT_FILE}" >&2
+  exit 0
+fi
+
 TMP_DIR=$(mktemp --directory chart-repo.XXXXX)
 # shellcheck disable=SC2064
 trap "rm -Rf '$(realpath "$TMP_DIR")'" EXIT
